+++ cmake_format/__init__.py
@@ -1,0 +1,5 @@
+"""
+Parse cmake listfiles and format them nicely
+"""
+
+VERSION = '0.3.6'